Moved gql and rpc to api folder.
