+++ src/api/index.ts
@@ -1,0 +1,7 @@
+import client from "./graphQL/gqlClient";
+
+const api = {
+  client
+}
+
+export default api;
