+++ src/hooks/useApi.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react'
-import ApiContext, { ApiContextProps } from '../context/ApiContext'
+import ApiContext, { ApiContextProps } from '../api/chainApi/ApiContext'
 
 export function useApi(): ApiContextProps {
   return useContext(ApiContext)
